centralize data formater in a service class
