+++ src/app/components/person/manage/manage-person.service.ts
@@ -101,4 +101,31 @@ export class ManagePersonService {
     this.response.message = event.message;
     this.response.status = event.status;
   }
+
+  toDate(dateStr): Date {
+    const [day, month, year] = dateStr.toString().split('/');
+    return new Date(year, month - 1, day);
+  }
+
+  formatDate(dateToFormt: string) {
+    const dateArray = dateToFormt.substr(0, 10).split('-');
+    const day = dateArray[2];
+    const month = dateArray[1];
+    const year = dateArray[0];
+    return `${day}/${month}/${year}`;
+  }
+
+  toDatabaseDate(dateStr): string {
+    const [day, month, year] = dateStr.toString().split('/');
+    return `${year}-${month}-${day}`;
+  }
+
+  validateDate(date: string): boolean {
+    console.log('validate date', date);
+    const dateRegex = /^([0-2][0-9]|(3)[0-1])(\/)(((0)[0-9])|((1)[0-2]))(\/)\d{4}$/;
+    if (!dateRegex.test(date)) {
+      return false;
+    }
+    return true;
+  }
 }
